Hydra.Chain.ChainState

Documentation

newtype ChainSlot

A generic description for a chain slot all implementations need to use.

Constructors

ChainSlot Natural

class (IsTx tx, Eq (ChainStateType tx), Show (ChainStateType tx), FromJSON (ChainStateType tx), ToJSON (ChainStateType tx), Arbitrary (ChainStateType tx)) => IsChainState tx where

Types that can be used on-chain by the Hydra protocol. XXX: Find a better name for this. Maybe IsChainTx or IsL1Tx?

Associated Types

type ChainStateType tx = c | c -> tx

Type of what to keep as L1 chain state. XXX: Why is this not always UTxOType?

Methods

chainStateSlot :: ChainStateType tx -> ChainSlot

Get the chain slot for a chain state. NOTE: For any sequence of a encountered, we assume monotonically increasing slots.