Hydra.Tx.Abort

Documentation

data AbortTxError

Constructors

OverlappingInputs

CannotFindHeadOutputToAbort

abortTx

Arguments

:: UTxO
Committed UTxOs to reimburse.

-> ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> (TxIn, TxOut CtxUTxO)
Everything needed to spend the Head state-machine output.

-> PlutusScript
Script for monetary policy to burn tokens

-> Map TxIn (TxOut CtxUTxO)
Data needed to spend the initial output sent to each party to the Head. Should contain the PT and is locked by initial script.

-> Map TxIn (TxOut CtxUTxO)
Data needed to spend commit outputs. Should contain the PT and is locked by commit script.

-> Either AbortTxError Tx

Create transaction which aborts a head by spending the Head output and all other "initial" outputs.