Hydra.Tx.Close

Documentation

data OpenThreadOutput

Representation of the Head output after a CollectCom transaction.

Constructors

OpenThreadOutput

Fields

openThreadUTxO :: (TxIn, TxOut CtxUTxO)

openContestationPeriod :: ContestationPeriod

openParties :: [Party]

closeTx

Arguments

:: ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> HeadId
Head identifier

-> SnapshotVersion
Last known version of the open head.

-> ConfirmedSnapshot Tx
Snapshot with instructions how to close the head.

-> SlotNo
Lower validity slot number, usually a current or quite recent slot number.

-> PointInTime
Upper validity slot and UTC time to compute the contestation deadline time.

-> OpenThreadOutput
Everything needed to spend the Head state-machine output.

-> Tx

Create a transaction closing a head with either the initial snapshot or with a multi-signed confirmed snapshot.