Hydra.Tx.CollectCom

Documentation

collectComTx

Arguments

:: NetworkId

-> ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> HeadId
Head identifier

-> HeadParameters
Parameters of the head to collect .

-> (TxIn, TxOut CtxUTxO)
Everything needed to spend the Head state-machine output.

-> Map TxIn (TxOut CtxUTxO)
Data needed to spend the commit output produced by each party. Should contain the PT and is locked by ν_commit script.

-> UTxO
UTxO to be used to collect. Should match whatever is recorded in the commit inputs.

-> Tx

Create a transaction collecting all "committed" utxo and opening a Head, i.e. driving the Head script state.