Hydra.Tx.Contest

Construction

data ClosedThreadOutput

Constructors

ClosedThreadOutput

Fields

closedThreadUTxO :: (TxIn, TxOut CtxUTxO)

closedParties :: [Party]

closedContestationDeadline :: POSIXTime

closedContesters :: [PubKeyHash]

contestTx

Arguments

:: ScriptRegistry
Published Hydra scripts to reference.

-> VerificationKey PaymentKey
Party who's authorizing this transaction

-> HeadId

-> ContestationPeriod

-> SnapshotVersion

-> Snapshot Tx
Contested snapshot number (i.e. the one we contest to)

-> MultiSignature (Snapshot Tx)
Multi-signature of the whole snapshot

-> PointInTime
Current slot and posix time to be used as the contestation time.

-> ClosedThreadOutput
Everything needed to spend the Head state-machine output.

-> IncrementalAction

-> Tx

Observation

data ContestObservation

Constructors

ContestObservation

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

contesters :: [PubKeyHash]

observeContestTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe ContestObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.