Hydra.Tx.Init

Construction

initTx

Arguments

:: NetworkId

-> TxIn
Seed input.

-> [OnChainId]
Verification key hashes of all participants.

-> HeadParameters

-> Tx

Create the init transaction from some HeadParameters and a single TxIn which will be used as unique parameter for minting NFTs.

mkHeadOutput :: NetworkId -> PolicyId -> TxOutDatum ctx -> TxOut ctx

mkHeadOutputInitial :: NetworkId -> TxIn -> HeadParameters -> TxOut CtxTx

mkInitialOutput :: NetworkId -> TxIn -> OnChainId -> TxOut CtxTx

Observation

data InitObservation

Data which can be observed from an initTx.

Constructors

InitObservation

Fields

headId :: HeadId

headSeed :: HeadSeed

headParameters :: HeadParameters

participants :: [OnChainId]

data NotAnInitReason

Constructors

NoHeadOutput

NotAHeadDatum

InvalidPartyInDatum

NoSTFound

NotAHeadPolicy

observeInitTx :: Tx -> Either NotAnInitReason InitObservation

Identify a init tx by checking the output value for holding tokens that are valid head tokens (checked by seed + policy).