Hydra.Tx.Observe

Description

Observe hydra transactions

Documentation

data HeadObservation

Generalised type for arbitrary Head observations on-chain.

Constructors

NoHeadTx

Init InitObservation

Abort AbortObservation

Commit CommitObservation

CollectCom CollectComObservation

Deposit DepositObservation

Recover RecoverObservation

Increment IncrementObservation

Decrement DecrementObservation

Close CloseObservation

Contest ContestObservation

Fanout FanoutObservation

observeHeadTx :: NetworkId -> UTxO -> Tx -> HeadObservation

Observe any Hydra head transaction.

data InitObservation

Data which can be observed from an initTx.

Constructors

InitObservation

Fields

headId :: HeadId

headSeed :: HeadSeed

headParameters :: HeadParameters

participants :: [OnChainId]

data NotAnInitReason

Constructors

NoHeadOutput

NotAHeadDatum

InvalidPartyInDatum

NoSTFound

NotAHeadPolicy

observeInitTx :: Tx -> Either NotAnInitReason InitObservation

Identify a init tx by checking the output value for holding tokens that are valid head tokens (checked by seed + policy).

newtype AbortObservation

Constructors

AbortObservation

Fields

headId :: HeadId

observeAbortTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe AbortObservation

Identify an abort tx by looking up the input spending the Head output and decoding its redeemer.

data CommitObservation

Full observation of a commit transaction.

Constructors

CommitObservation

Fields

party :: Party

Hydra participant who committed the UTxO.

committed :: UTxO

headId :: HeadId

observeCommitTx

Arguments

:: NetworkId

-> UTxO
A UTxO set to lookup tx inputs. Should at least contain the input spending from νInitial.

-> Tx

-> Maybe CommitObservation

Identify a commit tx by:

Check that its spending from the init validator,

Find the outputs which pays to the commit validator,

Using the datum of that output, deserialize the committed output,

Reconstruct the committed UTxO from both values (tx input and output).

data CollectComObservation

Constructors

CollectComObservation

Fields

headId :: HeadId

utxoHash :: UTxOHash

observeCollectComTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CollectComObservation

Identify a collectCom tx by lookup up the input spending the Head output and decoding its redeemer.

data DecrementObservation

Constructors

DecrementObservation

Fields

headId :: HeadId

newVersion :: SnapshotVersion

distributedUTxO :: UTxO

observeDecrementTx :: UTxO -> Tx -> Maybe DecrementObservation

data DepositObservation

Constructors

DepositObservation

Fields

headId :: HeadId

depositTxId :: TxId

deposited :: UTxO

created :: SlotNo

deadline :: UTCTime

observeDepositTx :: NetworkId -> Tx -> Maybe DepositObservation

Observe a deposit transaction by decoding the target head id, deposit deadline and deposited utxo in the datum.

This includes checking whether - the first output is a deposit output - all inputs of deposited utxo are actually spent, - the deposit script output actually contains the deposited value, - an upper validity bound has been set (used as creation slot).

data IncrementObservation

Constructors

IncrementObservation

Fields

headId :: HeadId

newVersion :: SnapshotVersion

depositTxId :: TxId

observeIncrementTx :: UTxO -> Tx -> Maybe IncrementObservation

data RecoverObservation

Constructors

RecoverObservation

Fields

headId :: HeadId

recoveredTxId :: TxId

recoveredUTxO :: UTxO

observeRecoverTx :: NetworkId -> UTxO -> Tx -> Maybe RecoverObservation

data CloseObservation

Constructors

CloseObservation

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

observeCloseTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe CloseObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

data ContestObservation

Constructors

ContestObservation

Fields

headId :: HeadId

snapshotNumber :: SnapshotNumber

contestationDeadline :: UTCTime

contesters :: [PubKeyHash]

observeContestTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe ContestObservation

Identify a close tx by lookup up the input spending the Head output and decoding its redeemer.

data FanoutObservation

Constructors

FanoutObservation

Fields

headId :: HeadId

fanoutUTxO :: UTxO

observeFanoutTx

Arguments

:: UTxO
A UTxO set to lookup tx inputs

-> Tx

-> Maybe FanoutObservation

Identify a fanout tx by lookup up the input spending the Head output and decoding its redeemer.