Hydra.Tx.Utils

Documentation

hydraHeadV1AssetName :: AssetName

hydraMetadataLabel :: Word64

The metadata label used for identifying Hydra protocol transactions. As suggested by a friendly large language model: The number most commonly associated with Hydra is 5, as in the mythological creature Hydra, which had multiple heads, and the number 5 often symbolizes multiplicity or diversity. However, there is no specific numerical association for Hydra smaller than 10000 beyond this mythological reference.

mkHydraHeadV1TxName :: Text -> TxMetadata

Create a transaction metadata entry to identify Hydra transactions (for informational purposes).

assetNameToOnChainId :: AssetName -> OnChainId

onChainIdToAssetName :: OnChainId -> AssetName

findFirst :: Foldable t => (a -> Maybe b) -> t a -> Maybe b

Find first occurrence including a transformation.

verificationKeyToOnChainId :: VerificationKey PaymentKey -> OnChainId

Derive the OnChainId from a Cardano PaymentKey. The on-chain identifier is the public key hash as it is also availble to plutus validators.

headTokensFromValue :: PlutusScript -> Value -> [(AssetName, Quantity)]

splitUTxO :: UTxO -> (UTxO, UTxO)

Split a given UTxO into two, such that the second UTxO is non-empty. This is useful to pick a UTxO to decommit.

adaOnly :: TxOut CtxUTxO -> TxOut CtxUTxO

extractInlineDatumFromTxOut :: FromData a => TxOut CtxUTxO -> Maybe a

Extract the inline datum from a given TxOut.