Hydra.Tx.Deposit

Construction

depositTx

Arguments

:: NetworkId

-> PParams LedgerEra

-> HeadId

-> CommitBlueprintTx Tx

-> SlotNo
Slot to use as upper validity. Will mark the time of creation of the deposit.

-> UTCTime
Deposit deadline from which onward the deposit can be recovered.

-> Maybe AddressInEra

-> Tx

Builds a deposit transaction to lock funds into the v_deposit script.

mkDepositOutput :: NetworkId -> HeadId -> UTxO -> UTCTime -> TxOut ctx

constructDepositUTxO :: TxId -> [TxOut CtxTx] -> UTxO

depositAddress :: NetworkId -> AddressInEra

Observation

data DepositObservation

Constructors

DepositObservation

Fields

headId :: HeadId

depositTxId :: TxId

deposited :: UTxO

created :: SlotNo

deadline :: UTCTime

observeDepositTx :: NetworkId -> Tx -> Maybe DepositObservation

Observe a deposit transaction by decoding the target head id, deposit deadline and deposited utxo in the datum.

This includes checking whether - the first output is a deposit output - all of deposited value is contained in the deposit tx output, - the deposit script output actually contains the deposited value, - an upper validity bound has been set (used as creation slot).

observeDepositTxOut :: Network -> TxOut CtxUTxO -> Maybe (HeadId, UTxO, POSIXTime)