Hydra.Tx.Utils

Documentation

data IncrementalAction

Type to encapsulate one of the two possible incremental actions or a regular snapshot. This actually signals that our snapshot modeling is likely not ideal but for now we want to keep track of both fields (de/commit) since we might want to support batch de/commits too in the future, but having both fields be Maybe UTxO intruduces a lot of checks if the value is Nothing or mempty.

Constructors

ToCommit UTxO

ToDecommit UTxO

NoThing

findStateToken :: TxOut ctx -> Maybe HeadId

Find (if it exists) the head identifier contained in given TxOut.

headTokensFromValue :: PlutusScript -> Value -> PolicyAssets

mkHydraHeadV1TxName :: Text -> TxMetadata

Create a transaction metadata entry to identify Hydra transactions (for informational purposes).

addMetadata :: TxMetadata -> Tx -> AlonzoTx ConwayEra -> AlonzoTx ConwayEra

assetNameToOnChainId :: AssetName -> OnChainId

findFirst :: Foldable t => (a -> Maybe b) -> t a -> Maybe b

Find first occurrence including a transformation.

hydraHeadV1AssetName :: AssetName

onChainIdToAssetName :: OnChainId -> AssetName

hydraMetadataLabel :: Word64

The metadata label used for identifying Hydra protocol transactions. As suggested by a friendly large language model: The number most commonly associated with Hydra is 5, as in the mythological creature Hydra, which had multiple heads, and the number 5 often symbolizes multiplicity or diversity. However, there is no specific numerical association for Hydra smaller than 10000 beyond this mythological reference.

verificationKeyToOnChainId :: VerificationKey PaymentKey -> OnChainId

Derive the OnChainId from a Cardano PaymentKey. The on-chain identifier is the public key hash as it is also availble to plutus validators.

splitUTxO :: UTxO -> (UTxO, UTxO)

Split a given UTxO into two, such that the second UTxO is non-empty. This is useful to pick a UTxO to decommit.

adaOnly :: TxOut CtxUTxO -> TxOut CtxUTxO

setIncrementalActionMaybe :: Maybe UTxO -> Maybe UTxO -> Maybe IncrementalAction

findHeadAssetId :: TxOut ctx -> Maybe (PolicyId, AssetName)

dummyValidatorScript :: PlutusScript